src/ZooCorp.BusinessLogic/Animals/Birds/Parrot.cs
using System.Collections.Generic;
using ZooCorp.BusinessLogic.Common;

namespace ZooCorp.BusinessLogic.Animals.Birds
{
    public class Parrot : Bird
    {
        public override int RequiredSpaceSfFt { get; } = 5;

        public override string[] FavoriteFood { get; } = new[] {"Vegetable"};

        public override List<string> FriendlyWith { get; } = new List<string>()
        {
            "Parrot",
            "Bison",
            "Elephant",
            "Turtle"
        };

        public Parrot(int id, List<int> feedSchedule = null, IConsole console = null) : base(console)
        {
            ID = id;
            FeedSchedule = feedSchedule ?? new List<int>() {6, 17};
        }

        public override bool IsFriendlyWith(Animal animal)
        {
            return FriendlyWith.Contains(animal.GetType().Name);
        }
    }
}